LayoutTests/fast/mediastream/constructors-expected.txt

Tests the JSEP PeerConnection related constructors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof webkitPeerConnection00 === 'function' is true
PASS typeof SessionDescription === 'function' is true
PASS typeof IceCandidate === 'function' is true
PASS typeof webkitRTCPeerConnection === 'function' is true
PASS typeof RTCSessionDescription === 'function' is true
PASS typeof RTCIceCandidate === 'function' is true
PASS webkitPeerConnection00() threw exception TypeError: DOM object constructor cannot be called as a function..
PASS SessionDescription() threw exception TypeError: DOM object constructor cannot be called as a function..
PASS IceCandidate() threw exception TypeError: DOM object constructor cannot be called as a function..
PASS webkitRTCPeerConnection() threw exception TypeError: DOM object constructor cannot be called as a function..
PASS RTCSessionDescription() threw exception TypeError: DOM object constructor cannot be called as a function..
PASS RTCIceCandidate() threw exception TypeError: DOM object constructor cannot be called as a function..
PASS new webkitPeerConnection00('STUN foobar.com:12345', function(){}); did not throw exception.
PASS new SessionDescription(''); did not throw exception.
PASS new IceCandidate('', ''); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{url:'stun://foobar.com:12345'}]}, null); did not throw exception.
PASS new RTCSessionDescription({type:'offer',sdp:'foobar'}); did not throw exception.
PASS new RTCIceCandidate({candidate:'foo'}); did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
